+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.3.1 (2022-07-07)
+## 0.3.2 (2023-16-11)
+
+* Added KnotBindingsProperty constructor
+* Added KnotBindingsProperty.GetFromSetterPriority method
+
+## 0.3.1 (2023-07-07)
 
 * Added AnyPropertyChanged event to KnotBindingsContainer